requirements.txt
openai
openai-whisper
torch>=2.0.0
pyaudio
pynput
keyboard
argparse
rumps
numpy
